pyproject.toml
[project]
name = "m365-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.14.5",
    "httpx>=0.28.1",
    "msal>=1.34.0",
    "python-dotenv>=1.2.1",
]
[project.scripts]
m365-assistant = "m365_assistant.main:run"